// frontend/src/app/shared/generated-api/model/adminUserUpdate.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdminUserUpdate { 
    username?: string;
    email?: string;
    displayName?: string;
    role?: AdminUserUpdateRoleEnum;
}
export enum AdminUserUpdateRoleEnum {
    User = 'USER',
    Admin = 'ADMIN'
};