// frontend/src/app/shared/generated-api/model/importApplyRequest.ts
/**
 * Glacier Notes Cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ImportApplyRequest { 
    strategy: ImportApplyRequestStrategyEnum;
}
export enum ImportApplyRequestStrategyEnum {
    Preserve = 'PRESERVE',
    AddAsCopies = 'ADD_AS_COPIES',
    ReplaceById = 'REPLACE_BY_ID'
};